refactor: rename projects to repositories
