+++ src/components/RepositorySection.tsx
@@ -5,18 +5,18 @@ import Link from 'next/link';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import projects from '../projects.json';
-import { Button, ProjectCard } from './common';
+import projects from '../repositories.json';
+import { Button, RepositoryCard } from './common';
 
 interface IProps { }
 
-export const ProjectSection: React.FC<IProps> = () => {
+export const RepositorySection: React.FC<IProps> = () => {
   const { t } = useTranslation();
   return (
-    <section className="project-section">
-      <div className="project-main">
-        <Image className="project-main-image" src="https://github.com/BRL-CAD.png" height={500} width={500} layout="fixed" />
-        <div className="project-main-body">
+    <section className="repository-section">
+      <div className="repository-main">
+        <Image className="repository-main-image" src="https://github.com/BRL-CAD.png" height={500} width={500} layout="fixed" />
+        <div className="repository-main-body">
           <h1>{t('PROJECT.TITLE')}</h1>
           <p>
             {t('PROJECT.DESCRIPTION')}
@@ -28,9 +28,9 @@ export const ProjectSection: React.FC<IProps> = () => {
           </Link>
         </div>
       </div>
-      <div className="project-section-list">
+      <div className="repository-section-list">
         {projects.slice(0, 2).map(({ title, description, link }) => (
-          <ProjectCard
+          <RepositoryCard
             key={title}
             title={title}
             description={description}
@@ -41,7 +41,7 @@ export const ProjectSection: React.FC<IProps> = () => {
       <Button
         elementType="a"
         className="brl-button--small brl-button--reverse"
-        href="/projects"
+        href="/repositories"
       >
         {t('PROJECT.BUTTON')}
       </Button>
