+++ src/components/common/index.ts
@@ -5,4 +5,4 @@ export { FeatureBlock } from './FeatureBlock';
 export { HighlightBlock } from './HighlightBlock';
 export { Modal } from './Modal';
 export { PageHeader } from './PageHeader';
-export { ProjectCard } from './ProjectCard';
+export { RepositoryCard } from './RepositoryCard';
