+++ src/i18n/locales/en.json
@@ -76,8 +76,8 @@
     }
   },
   "PROJECT": {
-    "TITLE": "Our Projects",
-    "DESCRIPTION": "BRL-CAD is a powerful cross-platform open source combinatorial solid modeling system that includes an interactive 3D solid geometry editor, a network-distributed symmetric multiprocessing (SMP) high-performance ray-tracer with support for rendering and geometric analysis, image and signal-processing tools, a system performance analysis benchmark suite, a flexible geometry scripting interface, and a high-performance geometric representation and analysis library.",
+    "TITLE": "Our Repositories",
+    "DESCRIPTION": "Mike Muuss began the initial architecture and design of BRL-CAD back in 1979. Development as a unified package began in 1983. The first public release was made in 1984. BRL-CAD became an open source project on 21 December 2004.",
     "BUTTON": "View All",
     "VIEW_ALL": "Visit Github"
   },
