+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import Head from 'next/head';
 import { useTranslation } from 'react-i18next';
 
 import {
-  AboutUs, Features, Highlights, Information, Main, ProjectSection,
+  AboutUs, Features, Highlights, Information, Main, RepositorySection,
 } from '../components';
 
 const Home: NextPage = () => {
@@ -19,7 +19,7 @@ const Home: NextPage = () => {
       <Information />
       <Highlights />
       <AboutUs />
-      <ProjectSection />
+      <RepositorySection />
     </>
   );
 };
