+++ src/styles/common/index.scss
@@ -3,5 +3,5 @@
 @import "./highlightBlock.scss";
 @import "./aboutBlock.scss";
 @import "./modal.scss";
-@import "./projectCard.scss";
+@import "./repositoryCard.scss";
 @import "./pageHeader.scss";
