+++ src/styles/common/repositoryCard.scss
@@ -1,4 +1,4 @@
-.project-card {
+.repository-card {
   @include flex();
   background: var(--white);
   padding: $rhythm * 2;
