+++ src/styles/index.scss
@@ -3,7 +3,7 @@
 @import "./media.scss";
 @import "./globals.scss";
 
-@import "./projectsPage.scss";
+@import "./repositoryPage.scss";
 
 @import "./common";
 @import "./aboutUs.scss";
@@ -12,7 +12,7 @@
 @import "./features.scss";
 @import "./information.scss";
 @import "./highlights.scss";
-@import "./projectSection.scss";
+@import "./repositorySection.scss";
 @import "./footer.scss";
 
 html {
