+++ src/styles/repositoryPage.scss
@@ -1,4 +1,4 @@
-.projects-body {
+.repository-body {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   row-gap: $rhythm * 3;
